requirements.txt
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
grpcio-status==1.41.1
httplib2==0.20.2
icmplib==3.0.1
idna==3.3
importlib-metadata==4.8.2
ipaddress==1.0.23
itsdangerous==2.0.1
Jinja2==3.0.3
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauth2client==4.1.3
packaging==21.2
proto-plus==1.19.7
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.4.0
pyparsing==2.4.7
pyrqlite===HEAD
python-nmap==0.7.1
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
